fixed compile warnings in bvh_builder_sah_mb

+++ kernels/bvh/bvh_builder_sah_mb.cpp
@@ -452,17 +452,19 @@ namespace embree
         pstate.init(iter,size_t(1024));
 
         /* iterate over all meshes in the scene */
-        PrimInfo pinfo = parallel_for_for_prefix_sum0( pstate, iter, PrimInfo(empty), [&](GridMesh* mesh, const range<size_t>& r, size_t k, unsigned int geomID) -> PrimInfo
-                                                       {
-                                                PrimInfo pinfo(empty);
-                                                for (size_t j=r.begin(); j<r.end(); j++)
-                                                {
-                                                  if (!mesh->valid(j,range<size_t>(0,1))) continue;
-                                                  BBox3fa bounds = empty;                                                  const PrimRef prim(bounds,geomID,unsigned(j));                                                             
-                                                  pinfo.add_center2(prim,mesh->getNumSubGrids(j));
-                                                }
-                                                return pinfo;
-                                              }, [](const PrimInfo& a, const PrimInfo& b) -> PrimInfo { return PrimInfo::merge(a,b); });
+        PrimInfo pinfo = parallel_for_for_prefix_sum0( pstate, iter, PrimInfo(empty), [&](GridMesh* mesh, const range<size_t>& r, size_t k, size_t geomID) -> PrimInfo {
+            
+            PrimInfo pinfo(empty);
+            for (size_t j=r.begin(); j<r.end(); j++)
+            {
+              if (!mesh->valid(j,range<size_t>(0,1))) continue;
+              BBox3fa bounds = empty;
+              const PrimRef prim(bounds,unsigned(geomID),unsigned(j));
+              pinfo.add_center2(prim,mesh->getNumSubGrids(j));
+            }
+            return pinfo;
+          }, [](const PrimInfo& a, const PrimInfo& b) -> PrimInfo { return PrimInfo::merge(a,b); });
+        
         size_t numPrimitives = pinfo.size();
         if (numPrimitives == 0) return pinfo;
 
@@ -471,29 +473,30 @@ namespace embree
         prims.resize(numPrimitives); 
 
         /* second run to fill primrefs and SubGridBuildData arrays */
-        pinfo = parallel_for_for_prefix_sum1( pstate, iter, PrimInfo(empty), [&](GridMesh* mesh, const range<size_t>& r, size_t k, unsigned int geomID, const PrimInfo& base) -> PrimInfo
-                                              {
-                                                k = base.size();
-                                                size_t p_index = k;
-                                                PrimInfo pinfo(empty);
-                                                for (size_t j=r.begin(); j<r.end(); j++)
-                                                {
-                                                  const GridMesh::Grid &g = mesh->grid(j);
-                                                  if (!mesh->valid(j,range<size_t>(0,1))) continue;
-
-                                                  for (unsigned int y=0; y<g.resY-1u; y+=2)
-                                                    for (unsigned int x=0; x<g.resX-1u; x+=2)
-                                                    {
-                                                      BBox3fa bounds = empty;
-                                                      if (!mesh->buildBounds(g,x,y,itime,bounds)) continue; // get bounds of subgrid
-                                                      const PrimRef prim(bounds,geomID,unsigned(p_index));
-                                                      pinfo.add_center2(prim);
-                                                      sgrids[p_index] = SubGridBuildData(x | g.get3x3FlagsX(x), y | g.get3x3FlagsY(y), unsigned(j));
+        pinfo = parallel_for_for_prefix_sum1( pstate, iter, PrimInfo(empty), [&](GridMesh* mesh, const range<size_t>& r, size_t k, size_t geomID, const PrimInfo& base) -> PrimInfo {
+            
+            k = base.size();
+            size_t p_index = k;
+            PrimInfo pinfo(empty);
+            for (size_t j=r.begin(); j<r.end(); j++)
+            {
+              const GridMesh::Grid &g = mesh->grid(j);
+              if (!mesh->valid(j,range<size_t>(0,1))) continue;
+              
+              for (unsigned int y=0; y<g.resY-1u; y+=2)
+                for (unsigned int x=0; x<g.resX-1u; x+=2)
+                {
+                  BBox3fa bounds = empty;
+                  if (!mesh->buildBounds(g,x,y,itime,bounds)) continue; // get bounds of subgrid
+                  const PrimRef prim(bounds,unsigned(geomID),unsigned(p_index));
+                  pinfo.add_center2(prim);
+                  sgrids[p_index] = SubGridBuildData(x | g.get3x3FlagsX(x), y | g.get3x3FlagsY(y), unsigned(j));
                                                       prims[p_index++] = prim;                
-                                                    }
-                                                }
-                                                return pinfo;
-                                              }, [](const PrimInfo& a, const PrimInfo& b) -> PrimInfo { return PrimInfo::merge(a,b); });
+                }
+            }
+            return pinfo;
+          }, [](const PrimInfo& a, const PrimInfo& b) -> PrimInfo { return PrimInfo::merge(a,b); });
+        
         assert(pinfo.size() == numPrimitives);
         return pinfo;
       }
@@ -506,18 +509,19 @@ namespace embree
 
         pstate.init(iter,size_t(1024));
         /* iterate over all meshes in the scene */
-        PrimInfoMB pinfoMB = parallel_for_for_prefix_sum0( pstate, iter, PrimInfoMB(empty), [&](GridMesh* mesh, const range<size_t>& r, size_t k, size_t /*geomID*/) -> PrimInfoMB
-                                                       {
-                                                         PrimInfoMB pinfoMB(empty);
-                                                         for (size_t j=r.begin(); j<r.end(); j++)
-                                                         {
-                                                           if (!mesh->valid(j, mesh->timeSegmentRange(t0t1))) continue;
-                                                           LBBox3fa bounds(empty);
-                                                           PrimInfoMB gridMB(0,mesh->getNumSubGrids(j));
-                                                           pinfoMB.merge(gridMB);
-                                                         }
-                                                         return pinfoMB;
-                                                       }, [](const PrimInfoMB& a, const PrimInfoMB& b) -> PrimInfoMB { return PrimInfoMB::merge2(a,b); });
+        PrimInfoMB pinfoMB = parallel_for_for_prefix_sum0( pstate, iter, PrimInfoMB(empty), [&](GridMesh* mesh, const range<size_t>& r, size_t k, size_t /*geomID*/) -> PrimInfoMB {
+            
+            PrimInfoMB pinfoMB(empty);
+            for (size_t j=r.begin(); j<r.end(); j++)
+            {
+              if (!mesh->valid(j, mesh->timeSegmentRange(t0t1))) continue;
+              LBBox3fa bounds(empty);
+              PrimInfoMB gridMB(0,mesh->getNumSubGrids(j));
+              pinfoMB.merge(gridMB);
+            }
+            return pinfoMB;
+          }, [](const PrimInfoMB& a, const PrimInfoMB& b) -> PrimInfoMB { return PrimInfoMB::merge2(a,b); });
+        
         size_t numPrimitives = pinfoMB.size();
         if (numPrimitives == 0) return pinfoMB;
 
@@ -525,27 +529,28 @@ namespace embree
         sgrids.resize(numPrimitives); 
         prims.resize(numPrimitives); 
         /* second run to fill primrefs and SubGridBuildData arrays */
-        pinfoMB = parallel_for_for_prefix_sum1( pstate, iter, PrimInfoMB(empty), [&](GridMesh* mesh, const range<size_t>& r, size_t k, unsigned int geomID, const PrimInfoMB& base) -> PrimInfoMB
-                                              {
-                                                k = base.size();
-                                                size_t p_index = k;
-                                                PrimInfoMB pinfoMB(empty);
-                                                for (size_t j=r.begin(); j<r.end(); j++)
-                                                {
-                                                  if (!mesh->valid(j, mesh->timeSegmentRange(t0t1))) continue;
-                                                  const GridMesh::Grid &g = mesh->grid(j);
-
-                                                  for (unsigned int y=0; y<g.resY-1u; y+=2)
-                                                    for (unsigned int x=0; x<g.resX-1u; x+=2)
-                                                    {
-                                                      const PrimRefMB prim(mesh->linearBounds(g,x,y,t0t1),mesh->numTimeSegments(),mesh->time_range,mesh->numTimeSegments(),geomID,unsigned(p_index));
-                                                      pinfoMB.add_primref(prim);
-                                                      sgrids[p_index] = SubGridBuildData(x | g.get3x3FlagsX(x), y | g.get3x3FlagsY(y), unsigned(j));
-                                                      prims[p_index++] = prim;                
-                                                    }
-                                                }
-                                                return pinfoMB;
-                                              }, [](const PrimInfoMB& a, const PrimInfoMB& b) -> PrimInfoMB { return PrimInfoMB::merge2(a,b); });
+        pinfoMB = parallel_for_for_prefix_sum1( pstate, iter, PrimInfoMB(empty), [&](GridMesh* mesh, const range<size_t>& r, size_t k, size_t geomID, const PrimInfoMB& base) -> PrimInfoMB {
+            
+            k = base.size();
+            size_t p_index = k;
+            PrimInfoMB pinfoMB(empty);
+            for (size_t j=r.begin(); j<r.end(); j++)
+            {
+              if (!mesh->valid(j, mesh->timeSegmentRange(t0t1))) continue;
+              const GridMesh::Grid &g = mesh->grid(j);
+              
+              for (unsigned int y=0; y<g.resY-1u; y+=2)
+                for (unsigned int x=0; x<g.resX-1u; x+=2)
+                {
+                  const PrimRefMB prim(mesh->linearBounds(g,x,y,t0t1),mesh->numTimeSegments(),mesh->time_range,mesh->numTimeSegments(),unsigned(geomID),unsigned(p_index));
+                  pinfoMB.add_primref(prim);
+                  sgrids[p_index] = SubGridBuildData(x | g.get3x3FlagsX(x), y | g.get3x3FlagsY(y), unsigned(j));
+                  prims[p_index++] = prim;                
+                }
+            }
+            return pinfoMB;
+          }, [](const PrimInfoMB& a, const PrimInfoMB& b) -> PrimInfoMB { return PrimInfoMB::merge2(a,b); });
+        
         assert(pinfoMB.size() == numPrimitives);
         pinfoMB.time_range = t0t1;
         return pinfoMB;
